requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.7
pycurl==7.43.0
PyVimeo==1.0.9
requests==2.20.1
six==1.10.0
tinydb==3.12.1
tuspy==0.2.1
urllib3==1.24.1
